refactor!: move interrupt.rs to format/
It's not part of avutil, but avformat
