+++ src/format/mod.rs
@@ -3,7 +3,7 @@ pub use crate::util::format::AlphaMode;
 pub use crate::util::format::{pixel, Pixel};
 pub use crate::util::format::{sample, Sample};
 
-use crate::util::interrupt;
+pub mod interrupt;
 
 pub mod stream;
