+++ src/util/mod.rs
@@ -5,7 +5,6 @@ pub mod dictionary;
 pub mod error;
 pub mod format;
 pub mod frame;
-pub mod interrupt;
 pub mod log;
 pub mod mathematics;
 pub mod media;
